[skrifa] cleanup glyf code structure
No functional changes. Just some much needed renaming and code shuffling.
